+++ forge/changelog.md
@@ -1,2 +1,2 @@
-- Added OpenAL sound events to voice chat API
-- Move Sound Physics Remastered integration to voice chat API
+- Removed stereo config option
+- Added audio type option